refactor: removed code from root init
