+++ pyproject.toml
@@ -12,7 +12,6 @@ dependencies = [
     "aiosqlite>=0.20.0",
     "emoji-country-flag>=2.0.1",
     "polib>=1.2.0",
-    "bs4>=0.0.2",
     "lxml>=5.2.2",
     "meval>=2.5",
     "pillow>=10.4.0",
