+++ src/korone/utils/caching.py
@@ -1,4 +1,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2024 Hitalo M. <https://github.com/HitaloM>
 
-__version__ = "1.0.6"
+from cashews import Cache
+
+cache = Cache()
+cache.setup("redis://localhost", client_side=True)
